+++ src/lib/libast/disc/sfdcseekable.c
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 #include	"sfdchdr.h"
@@ -40,7 +41,7 @@ static ssize_t skwrite(Sfio_t* f, const void* buf, size_t n, Sfdisc_t* disc)
 	NOT_USED(buf);
 	NOT_USED(n);
 	NOT_USED(disc);
-	return (ssize_t)(-1);
+	return -1;
 }
 
 static ssize_t skread(Sfio_t*	f,	/* stream involved */
@@ -60,14 +61,14 @@ static ssize_t skread(Sfio_t*	f,	/* stream involved */
 
 	addr = sfseek(sf,0,SEEK_CUR);
 
-	if(addr+n <= sk->extent)
+	if((addr+(ssize_t)n) <= sk->extent)
 		return sfread(sf,buf,n);
 
 	if((r = (ssize_t)(sk->extent-addr)) > 0)
-	{	if((w = sfread(sf,buf,r)) != r)
+	{	if((w = sfread(sf,buf,(size_t)r)) != r)
 			return w;
 		buf = (char*)buf + r;
-		n -= r;
+		n -= (size_t)r;
 	}
 
 	/* do a raw read */
@@ -77,7 +78,7 @@ static ssize_t skread(Sfio_t*	f,	/* stream involved */
 	}
 	else
 	{
-		if((p = sfwrite(sf,buf,w)) != w)
+		if((p = sfwrite(sf,buf,(size_t)w)) != w)
 			sk->eof = 1;
 		if(p > 0)
 			sk->extent += p;
@@ -119,12 +120,12 @@ static Sfoff_t skseek(Sfio_t* f, Sfoff_t addr, int type, Sfdisc_t* disc)
 
 		/* read enough to reach the seek point */
 		while(addr > sk->extent)
-		{	if(addr > sk->extent+sizeof(buf) )
+		{	if(addr > sk->extent+(ssize_t)sizeof(buf) )
 				w = sizeof(buf);
 			else	w = (int)(addr-sk->extent);
-			if((r = sfrd(f,buf,w,disc)) <= 0)
+			if((r = sfrd(f,buf,(size_t)w,disc)) <= 0)
 				w = r-1;
-			else if((w = sfwrite(sf,buf,r)) > 0)
+			else if((w = sfwrite(sf,buf,(size_t)r)) > 0)
 				sk->extent += w;
 			if(w != r)
 			{	sk->eof = 1;
