+++ src/lib/libast/tm/tvsleep.c
@@ -70,7 +70,7 @@ do_sleep(const Tv_t* tv, Tv_t* rv)
 	if ((r = nanosleep(&stv, &srv)) && errno == EINTR && rv)
 	{
 		rv->tv_sec = srv.tv_sec;
-		rv->tv_nsec = srv.tv_nsec;
+		rv->tv_nsec = (uint32_t)srv.tv_nsec;
 	}
 	return r;
 }
